Nix: Remove flake-utils from flake inputs
Motive: flake-utils is an unnecessary abstraction library that manages to
creep into every project under the guise of making things "easier." There
is no reason to use flake-utils here, as Nix is a powerful enough DSL to
handle system abstractions in-house without relying on more 3rd party
code, especially in a flake as small as Ladybird's.

We are not re-inventing the wheel; we are merely going back to circle
wheels.